1、蚁群算法中每个结点保留当前发现的最好解对应的取值（蚁群算法）； 2、记录每一轮的totalcost没必要求平均（蚁群算法）； 3、记录每一轮的totalcostt长度固定（蚁群算法）。

+++ DCOPSolver/src/com/cqu/settings/Settings.java
@@ -13,7 +13,7 @@ public class Settings {
 	private int maxDimensionInMBDPOP;
 	
 	//蚁群算法参数设置
-	private long MaxCycle;
+	private int MaxCycle;
 	private int countAnt;
 	private int alpha;
 	private int beta;
@@ -87,11 +87,11 @@ public class Settings {
 	
 	
 
-	public long getMaxCycle() {
+	public int getMaxCycle() {
 		return MaxCycle;
 	}
 
-	public void setMaxCycle(long maxCycle) {
+	public void setMaxCycle(int maxCycle) {
 		MaxCycle = maxCycle;
 	}
